[Fix] Fixes issues due to database reworking

+++ src/Coordination/teacher.py
@@ -1,13 +1,15 @@
 import discord
 
-from Utils.database import *
 import Utils.environment as env
+
+from Utils.database import *
 from Utils.errors import *
+from Utils.logging import log
 
 
 # region Assignments
 
-async def assign_teacher(interaction: discord.Interaction, teacher: discord.Member, real_name: str, subject: str | None = None, phonenumber: str | None = None, availability: str | None = None):
+async def assign_teacher(interaction: discord.Interaction, teacher: discord.Member, real_name: str, subjects: str | None = None, phonenumber: str | None = None, availability: str | None = None):
     if interaction.guild is None:
         raise CodeError("Dieser Befehl kann nur in einem Server verwendet werden")
 
@@ -18,11 +20,11 @@ async def assign_teacher(interaction: discord.Interaction, teacher: discord.Memb
 
     # Setup teacher in db
     db_teacher = Teacher(teacher.id)
-    db_teacher.edit(real_name=real_name, subject=subject, phonenumber=phonenumber, availability=availability)
+    db_teacher.edit(real_name=real_name, subjects=subjects, phonenumber=phonenumber, availability=availability)
 
     # Configure teachers category
     teacher_category_name = env.generate_member_nick(db_teacher)
-    teacher_category = env.__unwrapped_get(interaction.guild.categories, teacher_category_name)  # Search for teacher category, because maybe there exists one already
+    teacher_category = discord.utils.get(interaction.guild.categories, name=teacher_category_name)  # Search for existing category, because maybe there exists one already
     if not teacher_category:
         # Create new teacher category
         overwrites = {
@@ -31,6 +33,9 @@ async def assign_teacher(interaction: discord.Interaction, teacher: discord.Memb
         }
 
         teacher_category = await interaction.guild.create_category(teacher_category_name, overwrites=overwrites)
+
+    # Connect teacher category with teacher
+    db_teacher.edit(teaching_category=teacher_category.id)
 
     # Create cmd channel for the teacher
     cmd_channel = await interaction.guild.create_text_channel('cmd', category=teacher_category, overwrites=overwrites)
@@ -55,9 +60,13 @@ async def unassign_teacher(interaction: discord.Interaction, teacher: discord.Me
 
     # Ensure that teacher has no current students
     teacher_category_name = env.generate_member_nick(db_teacher)
-    teacher_category = env.__unwrapped_get(interaction.guild.categories, teacher_category_name)
-    if [channel.name for channel in teacher_category.text_channels if channel.name != 'cmd'] != []:
-        raise UsageError(f"{teacher.mention} hat noch registrierte Schüler")
+    teacher_category = discord.utils.get(interaction.guild.categories, name=teacher_category_name)
+
+    if teacher_category:
+        if [channel.name for channel in teacher_category.text_channels if channel.name != 'cmd'] != []:
+            raise UsageError(f"{teacher.mention} hat noch registrierte Schüler")
+    else:
+        await log(interaction.guild, f"Lehrer {teacher.mention} hat keine Kategorie", details={'Teacher': f'{teacher.mention}'})
 
     # Reset teacher in db
     db_teacher.pop()
@@ -67,9 +76,10 @@ async def unassign_teacher(interaction: discord.Interaction, teacher: discord.Me
     await teacher.edit(nick=None)
 
     # Delete channels and category
-    for channel in teacher_category.text_channels:
-        await channel.delete()
-    await teacher_category.delete()
+    if teacher_category:
+        for channel in teacher_category.text_channels:
+            await channel.delete()
+        await teacher_category.delete()
 
 # endregion
